# internal/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Basic authentication for OAuth2 client credentials",
        "title": "Easyflow OAuth2 Server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@easyflow.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/.well-known/jwks.json": {
            "get": {
                "description": "Returns the JSON Web Key Set (JWKS) for token verification following RFC 7517",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Well-Known"
                ],
                "summary": "Get JSON Web Key Set",
                "responses": {
                    "200": {
                        "description": "JSON Web Key Set",
                        "schema": {
                            "$ref": "#/definitions/internal_server_routes_wellknown.JWKSet"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/easyflow-oauth2-server_internal_shared_errors.APIError"
                        }
                    }
                }
            }
        },
        "/.well-known/oauth-authorization-server": {
            "get": {
                "description": "Returns OAuth 2.0 Authorization Server Metadata following RFC 8414",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Well-Known"
                ],
                "summary": "Get OAuth2 Authorization Server Metadata",
                "responses": {
                    "200": {
                        "description": "OAuth2 server metadata",
                        "schema": {
                            "$ref": "#/definitions/internal_server_routes_wellknown.OAuth2Metadata"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/easyflow-oauth2-server_internal_shared_errors.APIError"
                        }
                    }
                }
            }
        },
        "/.well-known/openid-configuration": {
            "get": {
                "description": "Returns OAuth 2.0 Authorization Server Metadata following RFC 8414",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Well-Known"
                ],
                "summary": "Get OAuth2 Authorization Server Metadata",
                "responses": {
                    "200": {
                        "description": "OAuth2 server metadata",
                        "schema": {
                            "$ref": "#/definitions/internal_server_routes_wellknown.OAuth2Metadata"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/easyflow-oauth2-server_internal_shared_errors.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "easyflow-oauth2-server_internal_shared_errors.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code represents the HTTP status code",
                    "type": "integer",
                    "example": 400
                },
                "details": {
                    "description": "Details contains additional error information (optional)"
                },
                "error": {
                    "description": "Error represents a predefined error code from the enum package",
                    "allOf": [
                        {
                            "$ref": "#/definitions/easyflow-oauth2-server_internal_shared_errors.ErrorCode"
                        }
                    ],
                    "example": "INVALID_REQUEST_BODY"
                }
            }
        },
        "easyflow-oauth2-server_internal_shared_errors.ErrorCode": {
            "type": "string",
            "enum": [
                "UNAUTHORIZED",
                "NOT_ALLOWED",
                "NOT_FOUND",
                "ALREADY_EXISTS",
                "INTERNAL_SERVER_ERROR",
                "MISSING_SESSION_TOKEN",
                "INVALID_SESSION_TOKEN",
                "MISSING_CLIENT_ID",
                "INVALID_CLIENT_ID",
                "MISSING_CLIENT_SECRET",
                "INVALID_CLIENT_SECRET",
                "MISSING_CODE_CHALLENGE",
                "MISSING_STATE",
                "INVALID_STATE",
                "MISSING_REDIRECT_URI",
                "INVALID_REDIRECT_URI",
                "INVALID_CONTENT_TYPE",
                "INVALID_REQUEST_BODY",
                "MISSING_GRANT_TYPE",
                "INVALID_GRANT_TYPE",
                "MISSING_CODE",
                "INVALID_CODE",
                "MISSING_CODE_VERIFIER",
                "INVALID_CODE_VERIFIER",
                "MISSING_REFRESH_TOKEN",
                "INVALID_REFRESH_TOKEN"
            ],
            "x-enum-varnames": [
                "Unauthorized",
                "NotAllowed",
                "NotFound",
                "AlreadyExists",
                "InternalServerError",
                "MissingSessionToken",
                "InvalidSessionToken",
                "MissingClientID",
                "InvalidClientID",
                "MissingClientSecret",
                "InvalidClientSecret",
                "MissingCodeChallenge",
                "MissingState",
                "InvalidState",
                "MissingRedirectURI",
                "InvalidRedirectURI",
                "InvalidContentType",
                "InvalidRequestBody",
                "MissingGrantType",
                "InvalidGrantType",
                "MissingCode",
                "InvalidCode",
                "MissingCodeVerifier",
                "InvalidCodeVerifier",
                "MissingRefreshToken",
                "InvalidRefreshToken"
            ]
        },
...
        "internal_server_routes_wellknown.JWK": {
            "type": "object",
            "properties": {
                "alg": {
                    "description": "Algorithm (e.g., \"EdDSA\")",
                    "type": "string",
                    "example": "EdDSA"
                },
                "crv": {
                    "description": "Curve (e.g., \"Ed25519\")",
                    "type": "string",
                    "example": "Ed25519"
                },
                "e": {
                    "description": "Exponent (for RSA)",
                    "type": "string"
                },
                "kid": {
                    "description": "Key ID",
                    "type": "string",
                    "example": "key-1"
                },
                "kty": {
                    "description": "Key Type (e.g., \"OKP\" for Octet string key pairs)",
                    "type": "string",
                    "example": "OKP"
                },
                "n": {
                    "description": "Modulus (for RSA)",
                    "type": "string"
                },
                "use": {
                    "description": "Public Key Use (e.g., \"sig\" for signature)",
                    "type": "string",
                    "example": "sig"
                },
                "x": {
                    "description": "X Coordinate (base64url encoded)",
                    "type": "string",
                    "example": "11qYAYKxCrfVS_7TyWQHOg7hcvPapiMlrwIaaPcHURo"
                }
            }
        },
        "internal_server_routes_wellknown.JWKSet": {
            "type": "object",
            "properties": {
                "keys": {
                    "description": "Array of JSON Web Keys",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_server_routes_wellknown.JWK"
                    }
                }
            }
        },
        "internal_server_routes_wellknown.OAuth2Metadata": {
            "type": "object",
            "properties": {
                "authorization_endpoint": {
                    "description": "Authorization endpoint URL",
                    "type": "string",
                    "example": "https://auth.easyflow.com/oauth/authorize"
                },
                "code_challenge_methods_supported": {
                    "description": "Supported PKCE code challenge methods",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "S256"
                    ]
                },
                "grant_types_supported": {
                    "description": "Supported OAuth2 grant types",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "authorization_code",
                        "refresh_token"
                    ]
                },
                "introspection_endpoint": {
                    "description": "Token introspection endpoint",
                    "type": "string",
                    "example": "https://auth.easyflow.com/oauth/introspect"
                },
                "introspection_endpoint_auth_methods_supported": {
                    "description": "Supported introspection endpoint auth methods",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "client_secret_basic"
                    ]
                },
                "issuer": {
                    "description": "OAuth2 issuer identifier",
                    "type": "string",
                    "example": "https://auth.easyflow.com"
                },
                "jwks_uri": {
                    "description": "JSON Web Key Set URI",
                    "type": "string",
                    "example": "https://auth.easyflow.com/.well-known/jwks.json"
                },
                "op_policy_uri": {
                    "description": "Operator policy URI",
                    "type": "string",
                    "example": "https://easyflow.com/policy"
                },
                "op_tos_uri": {
                    "description": "Operator terms of service URI",
                    "type": "string",
                    "example": "https://easyflow.com/tos"
                },
                "registration_endpoint": {
                    "description": "Dynamic client registration endpoint",
                    "type": "string",
                    "example": "https://auth.easyflow.com/oauth/register"
                },
                "response_modes_supported": {
                    "description": "Supported response modes",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "query",
                        "fragment"
                    ]
                },
                "response_types_supported": {
                    "description": "Supported OAuth2 response types",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "code"
                    ]
                },
                "revocation_endpoint": {
                    "description": "Token revocation endpoint",
                    "type": "string",
                    "example": "https://auth.easyflow.com/oauth/revoke"
                },
                "revocation_endpoint_auth_methods_supported": {
                    "description": "Supported revocation endpoint auth methods",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "client_secret_basic"
                    ]
                },
                "scopes_supported": {
                    "description": "Supported scopes",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "openid",
                        "profile",
                        "email"
                    ]
                },
                "service_documentation": {
                    "description": "Service documentation URL",
                    "type": "string",
                    "example": "https://docs.easyflow.com"
                },
                "subject_types_supported": {
                    "description": "Supported subject types",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "public"
                    ]
                },
                "token_endpoint": {
                    "description": "Token endpoint URL",
                    "type": "string",
                    "example": "https://auth.easyflow.com/oauth/token"
                },
                "token_endpoint_auth_methods_supported": {
                    "description": "Supported token endpoint authentication methods",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "client_secret_basic",
                        "client_secret_post"
                    ]
                },
                "token_endpoint_auth_signing_alg_values_supported": {
                    "description": "Supported signing algorithms for token endpoint auth",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "RS256",
                        "ES256"
                    ]
                },
                "ui_locales_supported": {
                    "description": "Supported UI locales",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "en-US",
                        "de-DE"
                    ]
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "SessionToken": {
            "description": "Session token for authenticated users",
            "type": "apiKey",
            "name": "session_token",
            "in": "cookie"
        }
    }
}